# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrackTruck Live API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/allowed-emails:
    get:
      operationId: listAllowedEmails
      tags: [admin]
      summary: List all allowed emails and their roles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of allowed emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AllowedEmail"
        "403":
          description: Forbidden - Super Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: addAllowedEmail
      tags: [admin]
      summary: Add a new email to the allowed list
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddAllowedEmailRequest"
      responses:
        "201":
          description: Email added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllowedEmail"
        "400":
          description: Invalid request
        "403":
          description: Forbidden - Super Admin only
  /admin/allowed-emails/{email}:
    delete:
      operationId: removeAllowedEmail
      tags: [admin]
      summary: Remove an email from the allowed list
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: Email removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden - Super Admin only or trying to remove root super admin
        "404":
          description: Email not found

    put:
      operationId: updateAllowedEmail
      tags: [admin]
      summary: Update settings for an allowed email
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAllowedEmailRequest"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllowedEmail"
        "403":
          description: Forbidden
        "404":
          description: Not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    AllowedEmail:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [super_admin, admin, user]
        isPaid:
          type: boolean
        routeLimit:
          type: integer
        usedRoutes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - isPaid
        - routeLimit
        - usedRoutes
        - createdAt
    AddAllowedEmailRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [super_admin, admin, user]
          default: user
        isPaid:
          type: boolean
          default: true
        routeLimit:
          type: integer
          default: 25
      required:
        - email
    UpdateAllowedEmailRequest:
      type: object
      properties:
        name:
          type: string
        isPaid:
          type: boolean
        routeLimit:
          type: integer
        usedRoutes:
          type: integer
        role:
          type: string
          enum: [super_admin, admin, user]
      required: []