# lib/api-zod/src/generated/types/addAllowedEmailRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrackTruck Live API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AddAllowedEmailRequestRole } from "./addAllowedEmailRequestRole";
import type { AddAllowedEmailRequestRouteLimit } from "./addAllowedEmailRequestRouteLimit";

export interface AddAllowedEmailRequest {
  email: string;
  role?: AddAllowedEmailRequestRole;
  isPaid?: boolean;
  routeLimit?: AddAllowedEmailRequestRouteLimit;
}