# lib/api-zod/src/generated/types/allowedEmail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrackTruck Live API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AllowedEmailRole } from "./allowedEmailRole";

export interface AllowedEmail {
  id: number;
  email: string;
  role: AllowedEmailRole;
  isPaid: boolean;
  routeLimit: number;
  usedRoutes: number;
  createdAt: Date;
}