# lib/api-zod/src/generated/types/updateAllowedEmailRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrackTruck Live API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAllowedEmailRequestRouteLimit } from "./updateAllowedEmailRequestRouteLimit";

export interface UpdateAllowedEmailRequest {
  isPaid?: boolean;
  routeLimit?: UpdateAllowedEmailRequestRouteLimit;
}